requirements.txt
# Core orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0

# Data validation
pydantic>=2.0.0

# Storage
supabase>=2.0.0

# UI
streamlit>=1.30.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
langchain-anthropic
python-dotenv
